Final Changes

+++ priorfood.php
@@ -1,6 +1,14 @@
 <?php
 require_once('database.php');
 require_once('session.php');
+if (($output = message()) !== null) {
+	echo '<center><a style="color: white; background-color: #002147; padding: 5px 10px; border-radius: 5px; text-decoration: none; display: inline-block; font-size: 25px;">'.$output.'</a><center>';
+}
+if(isset($_SESSION['username']) && $_SESSION['username'] !== ""){
+}else{
+  $_SESSION['message'] = "You must be logged in to access this page";
+  header("Location: https://turing.cs.olemiss.edu/~kcking2/SeniorProject/home.php");
+}
 $mysqli = Database::dbConnect();
 $mysqli -> setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
 ?>
@@ -15,10 +23,11 @@ if (($output = message()) !== null) {
 }
 ?>
 <?php
-$query = "select distinct firstname, lastname,reqid,userid,date,fulfilled from requests inner join externalusers on requests.userid=externalusers.id where id = 8;
+$userid = $_SESSION['userid'];
+$query = "select distinct firstname, lastname,reqid,userid,date,fulfilled from requests inner join externalusers on requests.userid=externalusers.id where id = ?;
 ";
 $stmt = $mysqli-> prepare($query);
-$stmt -> execute();
+$stmt -> execute([$userid]);
 ?>
 <link rel="stylesheet" href="home2.css">
 
@@ -46,7 +55,7 @@ while ($row = $stmt->fetch(PDO::FETCH_ASSOC)) {
  echo "<td>".$fulfill."</td>";
  echo "<td><a href='viewrequest.php?id=" . $row['reqid'] ."'style='color:blue'>View Request</a></td>";
  if($row['fulfilled'] === 0){
- echo "<td><a href='deletereq.php?id=" . $row['reqid'] ."'style='color:red'>X</a></td>";
+ echo "<td><a href='deletereq.php?id=" . $row['reqid'] ."' onclick=\"return confirm('Are you sure you want to delete your food request?');\"style='color:red'>X</a></td>";
  }else{
  echo "<td>Request Cannot Be Deleted</td>";
  }
